requirements.txt
aio-pika==8.3.0
fastapi==0.88.0
uvicorn==0.20.0
pydantic==1.10.4
pyyaml==6.0
python-dotenv
email-validator